Also use float for buffer length/index.

+++ src/buffer.ml
@@ -5,7 +5,7 @@ let toString buffer =
 
 type buffer_class
 external buffer_class : buffer_class = "Buffer" [@@bs.val]
-external alloc : buffer_class -> int -> t = "" [@@bs.send]
+external alloc : buffer_class -> float -> t = "" [@@bs.send]
 let alloc = alloc buffer_class
 
 external concat : buffer_class -> t array -> t = "" [@@bs.send]
@@ -15,10 +15,10 @@ external from : buffer_class -> string -> string Js.Nullable.t -> t = "" [@@bs.s
 let from ?encoding data =
   from buffer_class data (Js.Nullable.fromOption encoding)
 
-let get : t -> int -> int = [%raw fun buf idx ->
+let get : t -> float -> int = [%raw fun buf idx ->
   "return buf[idx];"
 ]
 
-let set : t -> int -> int -> unit = [%raw fun buf idx el ->
+let set : t -> float -> int -> unit = [%raw fun buf idx el ->
   "buf[idx] = el;"
 ]
